+++ src/appenders/line/vite.config.js
@@ -29,7 +29,16 @@ export default defineConfig(({ command, mode, isSsrBuild, isPreview }) => {
         name: "LineAppenders",
         entry: path.resolve(__dirname, "index.ts"),
         formats: ["es", "cjs", "umd"],
-        fileName: (format, entryName) => `${entryName}.${format}.js`,
+        fileName: "index",
+      },
+      rollupOptions: {
+        external: [ "@logback4js/core", "@line/bot-sdk" ],
+        output: {
+          globals: {
+            "@logback4js/core": "core",
+            "@line/bot-sdk": "line_bot-sdk",
+          }
+        }
       },
     },
   };