+++ src/appenders/slack/vite.config.js
@@ -29,7 +29,16 @@ export default defineConfig(({ command, mode, isSsrBuild, isPreview }) => {
         name: "SlackAppenders",
         entry: path.resolve(__dirname, "index.ts"),
         formats: ["es", "cjs", "umd"],
-        fileName: (format, entryName) => `${entryName}.${format}.js`,
+        fileName: "index",
+      },
+      rollupOptions: {
+        external: [ "@logback4js/core", "@slack/web-api" ],
+        output: {
+          globals: {
+            "@logback4js/core": "core",
+            "@slack/web-api": "slack_web-api",
+          }
+        }
       },
     },
   };