+++ src/core/vite.config.js
@@ -29,7 +29,7 @@ export default defineConfig(({ command, mode, isSsrBuild, isPreview }) => {
         name: "core",
         entry: path.resolve(__dirname, "index.ts"),
         formats: ['es', 'cjs', 'umd'],
-        fileName: (format, entryName) => `${entryName}.${format}.js`,
+        fileName: "index",
       },
     },
   };